createproduct filter pagination responsive

+++ controllers/categoryController.js
@@ -1,4 +1,5 @@
 const Category = require('../models/categoryModel');
+const Products = require('../models/productModel');
 
 const categoryCtrl = {
     getCategory: async (req, res) => {
@@ -30,6 +31,10 @@ const categoryCtrl = {
 
     deleteCategory: async(req, res)=>{
         try {
+            const products = await Products.findOne({category: req.params.id})
+            if(products) return res.status(400).json({
+                msg: "Please delete All the Product with this Categories"
+            })
             await Category.findByIdAndDelete(req.params.id)
 
             res.json({message: "Category is deleted"})
